requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
gspread==6.1.2
google-auth==2.37.0
tzdata==2025.2

